first commit of featurebranch to extract away classes
