+++ packaged_classes/__init__.py
@@ -1,0 +1,7 @@
+__all__ = ["Tracker","Trackertwo"]
+
+
+from .tracker import Tracker
+from .trackertwo import Trackertwo
+
+print("successfully imported two relevant classes")
